FIXES IDE PERFORMANCE BY CUTTING Geometry.kt INTO PIECES and adds ray-obb intersection
